Refactor save module into modular components
